fix(cws): refactor timezone handling by moving logic from tz module to util
