+++ taos-ws-sys/src/ws/mod.rs
@@ -30,7 +30,6 @@ pub mod stmt;
 pub mod stmt2;
 pub mod stub;
 pub mod tmq;
-mod tz;
 pub mod util;
 
 pub type TAOS = c_void;
@@ -436,7 +435,7 @@ fn taos_init_impl() -> Result<(), Box<dyn std::error::Error>> {
         return Err(TaosError::new(Code::FAILED, &err).into());
     }
 
-    tz::set_tz_env(config::timezone().as_str());
+    util::set_tz_env(config::timezone().as_str());
 
     let mut layers = Vec::new();
     let log_dir = config::log_dir();
